Moved kestrel specific features into transport abstractions
